+++ crates/node/src/commands/run.rs
@@ -1,20 +1,35 @@
 use std::path::PathBuf;
 
+use madara_runtime::SealingMode;
 use mc_data_availability::DaLayer;
 use sc_cli::{Result, RpcMethods, RunCmd, SubstrateCli};
 use sc_service::BasePath;
+use serde::{Deserialize, Serialize};
 
 use crate::cli::Cli;
 use crate::service;
 
 /// Available Sealing methods.
-#[derive(Debug, Copy, Clone, clap::ValueEnum, Default)]
+#[derive(Debug, Copy, Clone, clap::ValueEnum, Default, Serialize, Deserialize)]
 pub enum Sealing {
-    // Seal using rpc method.
+    /// Seal using rpc method.
     #[default]
     Manual,
-    // Seal when transaction is executed.
+    /// Seal when transaction is executed. This mode does not finalize blocks, if you want to
+    /// finalize blocks use `--sealing=instant-finality`.
     Instant,
+    /// Seal when transaction is executed with finalization.
+    InstantFinality,
+}
+
+impl From<Sealing> for SealingMode {
+    fn from(value: Sealing) -> Self {
+        match value {
+            Sealing::Manual => SealingMode::Manual,
+            Sealing::Instant => SealingMode::Instant { finalize: false },
+            Sealing::InstantFinality => SealingMode::Instant { finalize: true },
+        }
+    }
 }
 
 #[derive(Clone, Debug, clap::Args)]
@@ -71,10 +86,9 @@ pub fn run_node(mut cli: Cli) -> Result<()> {
             None
         }
     };
-    let sealing = cli.run.sealing;
-    let cache = cli.run.cache;
-
     runner.run_node_until_exit(|config| async move {
+        let sealing = cli.run.sealing.map(Into::into).unwrap_or_default();
+        let cache = cli.run.cache;
         service::new_full(config, sealing, da_config, cache).map_err(sc_cli::Error::Service)
     })
 }